+++ Gruntfile.js
@@ -1,24 +1,22 @@
 module.exports = function (grunt) {
 	// CSS files to be built (relative to less directory, no extension)
 	var cssObjs = [
-		'all',
+		'pre',
+		'core',
+		'critical',
+		'deprecated',
 		//'modules/pagination',
 		//'modules/infographics',
-		//'critical'
 	];
 
 	var jsCssObjs = [
 		//'js-css/band_imagery',
-		//'js-css/display-font',
-		//'js-css/events',
-		//'js-css/events-band',
-		//'js-css/extlatin',
-		//'js-css/formvalidator',
-		//'js-css/monthwidget',
-		//'js-css/notices',
+		'js-css/events-band',
+		'js-css/events',
+		'js-css/formvalidator',
+		'js-css/notices',
+		'js-css/monthwidget',
 		//'js-css/rsswidget',
-		//'js-css/script-font',
-		//'js-css/smallcaps',
 		//'js-css/unlalert',
 		//'plugins/qtip/wdn.qtip',
 		//'plugins/ui/css/jquery-ui-wdn'
@@ -32,6 +30,7 @@ module.exports = function (grunt) {
 			templateCss = templateDir + '/css',                   // wdn/templates_5.0/css
 			templateJs = templateDir + '/js',                     // wdn/templates_5.0/js
 			templateJsSrc = templateDir + '/js-src',              // wdn/templates_5.0/js-src
+			templateJsCss = templateJs + '/js-css',               // wdn/templates_5.0/js/js-css
 			builtJsDir = 'compressed',                            // compressed
 			buildJsDir = buildDir + '/' + builtJsDir,             // build/compressed - folder removed at end of grunt js task
 			templateCompileJs = templateJs + '/' + builtJsDir,    // wdn/templates_5.0/js/compressed
@@ -43,16 +42,17 @@ module.exports = function (grunt) {
 
 	var hereDir = './';
 
-	// files for keyword replacement (should match .gitattributes)
+	// files for keyword replacement (e.g. DEP_VERSION)(should match .gitattributes file)
 	var filterFiles = [
 		templateHtmlDir + '/*.dwt*',
-		templateIncludeDir + '/scriptsandstyles*.html',
-		templateIncludeDir + '/speedy_body_scripts.html'
+		templateIncludeDir + '/global/*.html',
+		templateIncludeDir + '/local/*.html',
 	];
 
 	// polyfill modules that need sync loading (should match scripts loaded in debug.js)
 	var polyfillMods = [
 		//'modernizr-wdn',
+		'mustard-initializer', // make sure that polyfill.io and other mustard are loaded first before other scripts
 		'ga',
 		'requireLib',
 		'wdn'
@@ -60,21 +60,16 @@ module.exports = function (grunt) {
 
 	// modules added here will be added to rjsConfig modules below
 	var wdnBuildPlugins = [
-		'dialog',
 		//'band_imagery',
 		'carousel',
-		//'display-font',
-		//'events-band',
-		//'events',
-		//'extlatin',
+		'events-band',
+		'events',
 		//'jqueryui',
 		//'mediaelement_wdn',
 		//'modal',
-		//'monthwidget',
-		// 'notice',
+		'monthwidget',
+		'notice',
 		//'rss_widget',
-		//'script-font',
-		//'smallcaps',
 		//'tooltip'
 	];
 
@@ -84,7 +79,7 @@ module.exports = function (grunt) {
 		'require-css/normalize',
 		'jquery',
 		'wdn',
-		// 'plugins/hoverIntent/jquery.hoverIntent'
+		'plugins/hoverIntent/jquery.hoverIntent'
 	];
 
 	/**
@@ -93,12 +88,11 @@ module.exports = function (grunt) {
 	 */
 	var syncJsIgnore = [
 		'!build.txt',
-		'!js-css/**',
+		//'!js-css/**',
 		'!analytics.*',
 		'!debug.*',
 		'!fontfaceobserver.*',
 		'!fontsloaded.*',
-		'!form_validation.*',
 		'!ga.*',
 		'!jquery.*',
 		'!legacy.*',
@@ -141,6 +135,7 @@ module.exports = function (grunt) {
 		},
 		modules: [
 			//specify modules to include their immediate and deep dependencies in the final module file
+			// pro: one less HTTP request, con: if other modules share the same dependency, then dependency is not reusable
 			{
 				name: 'all',
 				create: true,
@@ -199,17 +194,18 @@ module.exports = function (grunt) {
 
 	var scssAllFiles = {}; // contains file patterns of temp scss files to compile scss from
 	cssObjs.forEach(function(file) {
+  	if (file.startsWith('pre')) return; // exclude this from Sass files compiled to CSS
 		scssAllFiles[templateCss + '/' + file + '.css'] = templateScss + '/' + file + '.tmp.scss';
 	});
 
 	var scssJsFiles = {};
 	jsCssObjs.forEach(function(file) {
-		scssJsFiles[templateJs + '/' + file + '.css'] = templateJs + '/' + file + '.scss';
+		scssJsFiles[templateJs + '/' + file + '.css'] = templateJsSrc + '/' + file + '.scss';
 	});
 
-	// load all grunt tasks matching the ['grunt-*', '@*/grunt-*'] patterns
+	// load all grunt plugins matching the ['grunt-*', '@*/grunt-*'] patterns
 	require('load-grunt-tasks')(grunt);
-
+	var nodeSass = require('node-sass');
 	/**
 	 * Setting up grunt tasks
 	 */
@@ -233,34 +229,52 @@ module.exports = function (grunt) {
 		},
 
 		sass: {
-			all: {
-				files: scssAllFiles,
+			main: {
 				options: {
+					implementation: nodeSass,
 					sourceMap: true,
 					includePaths: [
 						__dirname+'/node_modules/modularscale-sass/stylesheets'
 					]
-				}
+				},
+				files: scssAllFiles
 			},
-			js: {
+			plugins: {
 				options: {
-					sourceMap: true
+					implementation: nodeSass,
+					sourceMap: true,
+					includePaths: [
+						__dirname+'/node_modules/modularscale-sass/stylesheets'
+					]
 				},
 				files: scssJsFiles
 			}
 		},
 
 		postcss: {
-			options: {
-				processors: [
-					require('postcss-normalize')({forceImport:true}),
-					require('autoprefixer')
-				],
-				map: true
-			},
-			dist: {
-				src: templateCss + '/*.css'
-			}
+  		main: {
+  			options: {
+  				processors: [
+            require('postcss-normalize')({allowDuplicates: true}),
+            require('autoprefixer'),
+            require('postcss-object-fit-images'),
+            require('cssnano')()
+  				],
+  				map: true
+  			},
+        src: templateCss + '/*.css'
+  		},
+  		plugins: {
+  			options: {
+  				processors: [
+  					require('autoprefixer'),
+  					require('postcss-object-fit-images'),
+  					require('cssnano')()
+  				],
+  				map: true
+  			},
+        src: templateJsCss + '/*.css'
+      }
 		},
 
 		//set babel preset in .babelrc file
@@ -299,6 +313,27 @@ module.exports = function (grunt) {
 					].concat(syncJsIgnore),
 					dest: templateCompileJs
 				}]
+			},
+			dcfCommonModules: {
+				files: [{
+					cwd: 'node_modules/dcf/assets/dist/js/app/postBabel/common',
+					src: ['**/*.js', '!**/*.min.js', '!**/*.babel.js'],
+					dest: templateJs
+				}]
+			},
+			dcfOptionalModules: {
+				files: [{
+					cwd: 'node_modules/dcf/assets/dist/js/app/postBabel/optional',
+					src: ['**/*.js'],
+					dest: templateJs
+				}]
+			},
+			dcfUnminifiedMustards: {
+				files: [{
+					cwd: 'node_modules/dcf/assets/dist/js/mustard',
+					src: ['**/*.js', '!**/*.min.js'],
+					dest: `${templateJs}/mustard`
+				}]
 			}
 		},
 
@@ -316,7 +351,7 @@ module.exports = function (grunt) {
 		clean: {
 			scss: Object.keys(scssGlobAllTmpFiles).map((fileName) => `${templateScss}/${fileName}`),
 			css: [templateCss].concat(Object.keys(scssJsFiles)),
-			js: [templateCompileJs, templateJs + '/*.js'],
+			js: [templateJs],
 			"js-build": [buildJsDir],
 			dist: [zipDir + '/*.zip', zipDir + '/*.gz']
 		},
@@ -326,7 +361,7 @@ module.exports = function (grunt) {
 				files: [{
 					expand: true,
 					cwd: templateJsSrc,
-					src: ['**', '!**.babel.js'],
+					src: ['**', '!**/*.babel.js'],
 					dest: templateJs,
 					filter: 'isFile'
 				}]
@@ -358,7 +393,7 @@ module.exports = function (grunt) {
 		},
 
 		concurrent: {
-			main: [['sassGlobber:all', 'sass:all'], 'postcss', 'js'],
+			main: ['css-main', 'js'],
 			dist: ['zip', 'archive']
 		},
 
@@ -396,18 +431,49 @@ module.exports = function (grunt) {
 
 		watch: {
 			sass: {
-				files: [templateScss + '/**/*.scss', templateJs + '/js-css/*.scss'],
-				tasks: ['sassGlobber:all','sass']
+				files: [templateScss + '/**/*.scss', templateJsSrc + '/js-css/*.scss'],
+				tasks: ['sassGlobber', 'sass', 'postcss']
 			},
 			js: {
 				files: [templateJs + '/**/*.js', '!' + templateCompileJs + '/**/*.js'],
 				tasks: ['js']
 			},
-//			includes: {
-//				files: [buildDir + '/**/*.html', templateScss + '/**/*.scss', templateJs + '/js-css/*.scss'],
-//				tasks: ['includes']
-//			}
-		}
+			includes: {
+				files: [buildDir + '/**/*.html', templateScss + '/**/*.scss', templateJs + '/js-css/*.scss'],
+				tasks: ['includes']
+			}
+		},
+
+		// https://github.com/filamentgroup/grunt-criticalcss
+		// possible tool to get critical css
+    // criticalcss: {
+    //   desktop: {
+    //     options: {
+    //       url: "http://localhost/wdntemplates/debug.shtml",
+    //       width: 1200,
+    //       height: 900,
+    //       outputfile: "./desktop-critical.scss",
+    //       filename: "/Library/WebServer/Documents/wdntemplates/wdn/templates_5.0/css/core.css",
+    //       buffer: 800*1024,
+    //       ignoreConsole: false,
+		// 			forceInclude: [],
+    //       restoreFontFaces: false
+    //     }
+    //   },
+    //   mobile: {
+    //     options: {
+    //       url: "http://localhost/wdntemplates/debug.shtml",
+    //       width: 400,
+    //       height: 900,
+    //       outputfile: "./mobile-critical.scss",
+    //       filename: "/Library/WebServer/Documents/wdntemplates/wdn/templates_5.0/css/core.css",
+    //       buffer: 800*1024,
+    //       ignoreConsole: false,
+    //       forceInclude: [],
+    //       restoreFontFaces: false
+    //     }
+    //   }
+    // }
 	});
 
 	// keyword replacement task: restore keywords
@@ -475,12 +541,25 @@ module.exports = function (grunt) {
 		});
 	});
 
-	// establish grunt default
-	grunt.registerTask('default', ['concurrent:main']);
+  // https://github.com/filamentgroup/grunt-criticalcss
+	// use if want to run  criticalcss
+	// npm install grunt-criticalcss --save-dev
+  //grunt.loadNpmTasks('grunt-criticalcss');
 
+
+	// use if want to run  criticalcss
+  //grunt.registerTask('default', ['sassGlobber', 'clean:js', 'css-main', 'js-main', 'criticalcss'])
+
+
+	grunt.registerTask('css-main', ['sassGlobber', 'sass:main', 'postcss:main']);
+	grunt.registerTask('css-plugins', ['sassGlobber', 'sass:plugins', 'postcss:plugins']);
+	grunt.registerTask('js-main', ['css-plugins', 'babel', 'copy:babelNoTranspile', 'sync:dcfCommonModules', 'sync:dcfOptionalModules', 'sync:dcfUnminifiedMustards', 'requirejs', 'sync:js', 'clean:js-build']);
+	grunt.registerTask('js', ['clean:js', 'css-plugins', 'babel', 'copy:babelNoTranspile', 'requirejs', 'sync:js', 'clean:js-build']);
+
+	// establish grunt composed tasks
+	// TODO check with Ryan if sassGlobber needs to be at the start of Grunt task
+	grunt.registerTask('default', ['sassGlobber', 'clean:js', 'css-main', 'js-main']);
 	// legacy targets from Makefile
 	grunt.registerTask('dist', ['default', 'filter-smudge', 'concurrent:dist']);
-	grunt.registerTask('all', ['default']);
-	grunt.registerTask('js', ['clean:js', 'sass:js', 'babel', 'copy:babelNoTranspile', 'requirejs', 'sync:js', 'clean:js-build']);
-	grunt.registerTask('css', ['sassGlobber:all','sass:all', 'postcss']);
+	grunt.registerTask('all', ['default']);  /** mark for deletion */
 };