+++ wdn/templates_4.1/scripts/main-execute-mods.js
@@ -1,7 +1,0 @@
-// these modules are needed for framework loading, but require no initialization
-define([
-    'modernizr',
-    'wdn_jquery',
-    'legacy',
-    'skipnav'
-], function() {});