+++ wdn/templates_4.1/scripts/modernizr.js
@@ -1,4 +1,0 @@
-define('modernizr', [], function () {
-	// Modernizr should be loaded synchronously like requireLib
-	return window.Modernizr; 
-});