+++ wdn/templates_5.0/js-src/legacy.js
@@ -1,0 +1,12 @@
+define(['jquery', 'wdn', 'require'], function($, WDN, require) {
+	"use strict";
+
+	var setup = function() {
+		if ($('ul.wdn_tabs').length) {
+			WDN.initializePlugin('tabs');
+		}
+	};
+	$(setup);
+
+	return setup;
+});