+++ wdn/templates_5.0/js-src/main-wdn-plugins.babel.js
@@ -1,9 +1,12 @@
 "use strict";
 
-define(['analytics', //this also loads the IDM plugin
+define([
+'analytics', //this also loads the IDM plugin
 'navigation', //TODO: do we need navigation js?
 'search', //TODO: add search
-'unlalert'
+'unlalert',
+'legacy',
+'hover_intent'
 //'images', //TODO: do we need 'images'?
 //'form_validation', //TODO: do we need to init form_validation here?
 //'wdn-ui' //TODO: do we need js?