+++ wdn/templates_5.0/js-src/main.babel.js
@@ -17,5 +17,9 @@ requirejs([
 	//The following plugins initialize on their own and do not need to be passed to the callback
 	'plugins/skipto/skipto.min.js'
 ], function(WDN, require) {
-	//TODO: add UNLchat
+    var unlchat_url = 'https://ucommchat.unl.edu/assets/js';
+    //#UNLCHAT_URL
+    WDN.loadJQuery(function() {
+        require([unlchat_url + '?for=client&version=' + WDN.getHTMLVersion()], function(){});
+    });
 });